refactor: extract sandbox command interpreters
