+++ crates/agent-engine/src/tools/sandbox/command_options.rs
@@ -1,0 +1,12 @@
+pub(super) fn exact_or_inline_option_with_value(arg: &str, short: &[&str], long: &[&str]) -> bool {
+    short
+        .iter()
+        .any(|flag| arg == *flag || arg.starts_with(flag))
+        || long
+            .iter()
+            .any(|flag| arg == *flag || arg.starts_with(&format!("{flag}=")))
+}
+
+pub(super) fn has_attached_option_value(arg: &str) -> bool {
+    arg.contains('=') || (arg.starts_with('-') && !arg.starts_with("--") && arg.len() > 2)
+}
